+++ src/main/java/ru/gadjini/telegram/smart/bot/commons/job/UploadJob.java
@@ -134,7 +134,7 @@ public class UploadJob extends WorkQueueJobPusher {
         deleteUploads(producer, producerIds);
     }
 
-    public void cancelUploadsByUserId(String producer, int userId) {
+    public void cancelUploadsByUserId(String producer, long userId) {
         List<UploadQueueItem> deleted = uploadQueueService.deleteAndGetProcessingOrWaitingByUserId(producer, userId);
         uploadTasksExecutor.cancel(deleted.stream().map(UploadQueueItem::getId).collect(Collectors.toList()), true);
         uploadQueueService.releaseResources(deleted);
